Resize Loading Screen

+++ MissionEcopossible/Assets/Scripts/LevelLoader.cs
@@ -43,7 +43,7 @@ public class LevelLoader : MonoBehaviour
             float progress = Mathf.Clamp01(asyncLoad.progress / .9f);
 
             slider.value = progress;
-            progressText.text = progress * 100f + "%";
+            progressText.text =(int) progress * 100f + "%";
 
             yield return null;
         }
